+++ app/services/pdf/beratungshilfe/beratungshilfe.pdf.ts
@@ -1,0 +1,143 @@
+import { CheckboxValue } from "~/components/inputs/Checkbox";
+import { BeratungshilfeFormularContext } from "~/models/flows/beratungshilfeFormular";
+
+export const staatlicheLeistungMapping = {
+  grundsicherung: "Grundsicherung",
+  asylbewerberleistungen: "Asylbewerberleistungen",
+  buergergeld: "Bürgergeld",
+  andereLeistung: "Andere Leistung",
+  keine: "Keine",
+};
+
+export const getOccupationDetails = (
+  context: BeratungshilfeFormularContext,
+  withAdditionalIncome = true,
+) => {
+  const description: string[] = [];
+
+  if (context.erwerbstaetig === "no") {
+    description.push("nicht erwerbstätig");
+  } else if (context.berufart) {
+    const occupation = "Erwerbstätig";
+    const occupationTypeSelected = getSelectedOptions(
+      {
+        selbststaendig: "selbstständig",
+        festangestellt: "festangestellt",
+      },
+      context.berufart,
+    );
+
+    description.push(
+      `${occupation}${
+        occupationTypeSelected ? " (" + occupationTypeSelected + ")" : ""
+      }`,
+    );
+  }
+
+  const berufsituationMapping = {
+    no: "",
+    pupil: "Schüler:in",
+    student: "Student:in",
+    retiree: "Rentner:in",
+  };
+
+  description.push(berufsituationMapping[context.berufsituation ?? "no"]);
+
+  if (context.weitereseinkommen && withAdditionalIncome) {
+    const otherIncomes = getSelectedOptions(
+      {
+        unterhaltszahlungen: "Unterhaltszahlungen",
+        wohngeld: "Wohngeld",
+        kindergeld: "Kindergeld",
+        bafoeg: "Bafög",
+        others: "Sonstiges",
+      },
+      context.weitereseinkommen,
+    );
+
+    description.push(otherIncomes);
+  }
+
+  return description.filter((value) => value).join(", ");
+};
+
+export const newPageHint = "Bitte im Anhang prüfen";
+
+export const isANewAttachmentPageNeeded = (
+  context: BeratungshilfeFormularContext,
+): Attachment => {
+  const descriptions = [];
+
+  if (context.bereich) {
+    // TODO move to another function and use strapi as a source
+    const bereichMapping = {
+      authorities: "Behörden",
+      living: "Wohnen",
+      work: "Arbeit",
+      separation: "Trennung & Unterhalt",
+      trade: "Handel & Verträge",
+      debt: "Schulden & Forderungen",
+      inheritance: "Erben",
+      criminalProcedure: "Strafverfahren",
+      other: "Sonstiges",
+    };
+
+    descriptions.push({
+      title: "Thema des Rechtsproblems:",
+      text: bereichMapping[context.bereich],
+    });
+  }
+
+  if (context.beschreibung) {
+    descriptions.push({
+      title: "Beschreibung Angelegenheit:",
+      text: context.beschreibung,
+    });
+  }
+
+  if (context.eigeninitiativeBeschreibung) {
+    descriptions.push({
+      title: "Eigenbemühungen:",
+      text: context.eigeninitiativeBeschreibung,
+    });
+  } else if (context.keineEigeninitiativeBeschreibung) {
+    descriptions.push({
+      title: "Keine Eigenbemühung, weil:",
+      text: context.keineEigeninitiativeBeschreibung,
+    });
+  }
+
+  if (context.sonstiges) {
+    descriptions.push({ title: "Weitere Anmerkung:", text: context.sonstiges });
+  }
+
+  return {
+    shouldCreateNewPage:
+      descriptions.map((x) => x.title + x.text).join(" ").length > 255,
+    descriptions,
+  };
+};
+
+export type Attachment = {
+  shouldCreateNewPage: boolean;
+  descriptions: { title: string; text: string }[];
+};
+
+export function getSelectedOptions(
+  mapping: { [key: string]: string },
+  options?: { [key: string]: CheckboxValue },
+) {
+  if (!options) {
+    return "";
+  }
+
+  return Object.entries(options)
+    .map(([key, value]) => {
+      if (value === CheckboxValue.on) {
+        return mapping[key];
+      }
+      return "";
+    })
+    .filter((entry) => entry)
+    .join(", ");
+}
