+++ app/services/pdf/beratungshilfe/sections/B_vorraussetzungen.ts
@@ -1,0 +1,17 @@
+import { BeratungshilfeFormularContext } from "~/models/flows/beratungshilfeFormular";
+import { BeratungshilfePDF } from "../beratungshilfe.generated";
+
+export function fillVorraussetzungen(
+  pdfFields: BeratungshilfePDF,
+  context: BeratungshilfeFormularContext,
+) {
+  pdfFields.bIndervorliegendenAngelegenheittrittkeineRechtsschutzversicherungein.value =
+    context.rechtsschutzversicherung === "no";
+  pdfFields.b3IndieserAngelegenheitistmirbisherBeratungshilfewederbewilligtnochversagtworden.value =
+    context.beratungshilfeBeantragt === "no";
+  pdfFields.b2IndieserAngelegenheitbestehtfurmichnachmeinerKenntniskeineandereMoeglichkeitkostenloseBeratungundVertretunginAnspruchzunehmen.value =
+    context.eigeninitiativeGrundvorraussetzung === "no";
+  pdfFields.b4IndieserAngelegenheitwirdoderwurdevonmirbisherkeingerichtlichesVerfahrengefuhrt.value =
+    context.klageEingereicht === "no";
+  pdfFields.c2Einkuenftenetto.value = context.einkommen;
+}
